requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
sqlalchemy
asyncpg